# replace powers list with reverse builder using jupyter notebook