--- a/.gitignore
+++ b/.gitignore
@@ -4,3 +4,4 @@ __pycache__/
 dist/
 *.egg-info/
 .pytest_cache/
+.ipynb_checkpoints/