--- a/shortscale.py
+++ b/shortscale.py
@@ -1,8 +1,7 @@
 """English conversion from number to string"""
 import sys
 
-__version__ = '1.1.1'
-
+__version__ = '1.2.0'
 
 def shortscale(num: int) -> str:
     words = ''