… tests/test_shortscale.py