Used jupyter notebook to work out simpler algorithm

![Screenshot 2023-02-23 at 10 52 25 PM](https://user-images.githubusercontent.com/849592/221048906-fef320bb-3274-4e1b-a208-53c103d66fcf.png)

On M1 local laptop, this implementation is faster

before
```
     50000 calls,    5000000 bytes,     2067 ns/call
    100000 calls,   10000000 bytes,     2072 ns/call
```

after
```
    100000 calls,   10000000 bytes,     1808 ns/call
    200000 calls,   20000000 bytes,     1810 ns/call
```